refactor(wallet): adopt per-controller directory layout for RemoteFeatureFlagController
Migrates the RemoteFeatureFlagController instance to the per-controller
directory convention (introduced by #8953, extended by #8977):
`instances/remote-feature-flag-controller/` now holds the config, the
colocated test, and a `RemoteFeatureFlagControllerInstanceOptions` type in
its own `types.ts`. `InstanceSpecificOptions` references that type instead of
an inline shape, and `instances/index.ts` + the CODEOWNERS `## Initialization`
entry use the directory form. No public exports or option shapes change.
