+++ packages/wallet/src/initialization/instances/index.ts
@@ -2,5 +2,5 @@ export { accountsController } from './accounts-controller/accounts-controller';
 export { approvalController } from './approval-controller/approval-controller';
 export { connectivityController } from './connectivity-controller/connectivity-controller';
 export { keyringController } from './keyring-controller/keyring-controller';
-export { remoteFeatureFlagController } from './remote-feature-flag-controller';
+export { remoteFeatureFlagController } from './remote-feature-flag-controller/remote-feature-flag-controller';
 export { storageService } from './storage-service/storage-service';
